+++ DTCD-StyleSystem/src/stories/BaseChip.stories.js
@@ -1,0 +1,38 @@
+import BaseChip from '../base-components/BaseChip/BaseChip';
+
+export default {
+  title: 'Example/BaseComponents/BaseChip',
+  argTypes: {
+    defaultSlot: { 
+      type: 'string',
+      description: 'Default slot',
+    },
+    close: { 
+      type: 'boolean',
+      description: 'Show or hide remove icon.',
+    },
+  },
+};
+
+const NAME_COMPONENT = 'base-chip';
+
+window.customElements.define(NAME_COMPONENT, BaseChip);
+
+const Template = (args) => {
+  const {
+    close,
+    defaultSlot,
+  } = args;
+
+  const chip = document.createElement(NAME_COMPONENT);
+  chip.close = close;
+  chip.innerHTML += defaultSlot;
+
+  return chip;
+};
+
+export const DefaultChip = Template.bind({});
+DefaultChip.args = {
+  defaultSlot: 'Tag',
+  close: false,
+};